Extract getAmountsQuery and getAmountsCall to helper file
