+++ src/entities/addLiquidity/addLiquidityV2/stable/addLiquidityStable.ts
@@ -2,23 +2,21 @@ import { encodeFunctionData } from 'viem';
 import { Token } from '@/entities/token';
 import { TokenAmount } from '@/entities/tokenAmount';
 import { StableEncoder } from '@/entities/encoders/stable';
-import { VAULT, MAX_UINT256, ZERO_ADDRESS } from '@/utils';
+import { VAULT, ZERO_ADDRESS } from '@/utils';
 import { vaultV2Abi } from '@/abi';
 import {
     AddLiquidityBase,
     AddLiquidityInput,
-    AddLiquidityKind,
     AddLiquidityBaseQueryOutput,
     AddLiquidityBuildCallOutput,
 } from '@/entities/addLiquidity/types';
-import { AddLiquidityAmounts, PoolState } from '@/entities/types';
+import { PoolState } from '@/entities/types';
 import {
     doAddLiquidityQuery,
-    getAmounts,
     getSortedTokens,
     parseAddLiquidityArgs,
 } from '@/entities/utils';
-import { getAmountsCall } from '../../helpers';
+import { getAmountsCall, getAmountsQuery } from '../../helpers';
 import { AddLiquidityV2BaseBuildCallInput } from '../types';
 import { getValue } from '@/entities/utils/getValue';
 
@@ -28,7 +26,7 @@ export class AddLiquidityStable implements AddLiquidityBase {
         poolState: PoolState,
     ): Promise<AddLiquidityBaseQueryOutput> {
         const sortedTokens = getSortedTokens(poolState.tokens, input.chainId);
-        const amounts = this.getAmountsQuery(sortedTokens, input);
+        const amounts = getAmountsQuery(sortedTokens, input);
 
         const userData = StableEncoder.encodeAddLiquidityUserData(
             input.kind,
@@ -107,41 +105,5 @@ export class AddLiquidityStable implements AddLiquidityBase {
                 TokenAmount.fromRawAmount(a.token, amounts.maxAmountsIn[i]),
             ),
         };
-    }
-
-    private getAmountsQuery(
-        poolTokens: Token[],
-        input: AddLiquidityInput,
-    ): AddLiquidityAmounts {
-        switch (input.kind) {
-            case AddLiquidityKind.Unbalanced: {
-                return {
-                    minimumBpt: 0n,
-                    maxAmountsIn: getAmounts(poolTokens, input.amountsIn),
-                    tokenInIndex: undefined,
-                };
-            }
-            case AddLiquidityKind.SingleToken: {
-                const tokenInIndex = poolTokens.findIndex((t) =>
-                    t.isSameAddress(input.tokenIn),
-                );
-                if (tokenInIndex === -1)
-                    throw Error("Can't find index of SingleToken");
-                const maxAmountsIn = Array(poolTokens.length).fill(0n);
-                maxAmountsIn[tokenInIndex] = MAX_UINT256;
-                return {
-                    minimumBpt: input.bptOut.rawAmount,
-                    maxAmountsIn,
-                    tokenInIndex,
-                };
-            }
-            case AddLiquidityKind.Proportional: {
-                return {
-                    minimumBpt: input.bptOut.rawAmount,
-                    maxAmountsIn: Array(poolTokens.length).fill(MAX_UINT256),
-                    tokenInIndex: undefined,
-                };
-            }
-        }
     }
 }
