+++ src/entities/addLiquidity/addLiquidityV2/weighted/addLiquidityWeighted.ts
@@ -2,22 +2,20 @@ import { encodeFunctionData } from 'viem';
 import { Token } from '@/entities/token';
 import { TokenAmount } from '@/entities/tokenAmount';
 import { WeightedEncoder } from '@/entities/encoders/weighted';
-import { VAULT, MAX_UINT256, ZERO_ADDRESS } from '@/utils';
+import { VAULT, ZERO_ADDRESS } from '@/utils';
 import { vaultV2Abi } from '@/abi';
 import {
     AddLiquidityBase,
     AddLiquidityBuildCallOutput,
     AddLiquidityInput,
-    AddLiquidityKind,
 } from '@/entities/addLiquidity/types';
-import { AddLiquidityAmounts, PoolState } from '@/entities/types';
+import { PoolState } from '@/entities/types';
 import {
     doAddLiquidityQuery,
-    getAmounts,
     getSortedTokens,
     parseAddLiquidityArgs,
 } from '@/entities/utils';
-import { getAmountsCall } from '../../helpers';
+import { getAmountsCall, getAmountsQuery } from '../../helpers';
 import { getValue } from '../../../utils/getValue';
 import {
     AddLiquidityV2BaseBuildCallInput,
@@ -30,7 +28,7 @@ export class AddLiquidityWeighted implements AddLiquidityBase {
         poolState: PoolState,
     ): Promise<AddLiquidityV2BaseQueryOutput> {
         const sortedTokens = getSortedTokens(poolState.tokens, input.chainId);
-        const amounts = this.getAmountsQuery(sortedTokens, input);
+        const amounts = getAmountsQuery(sortedTokens, input);
 
         const userData = WeightedEncoder.encodeAddLiquidityUserData(
             input.kind,
@@ -109,41 +107,5 @@ export class AddLiquidityWeighted implements AddLiquidityBase {
                 TokenAmount.fromRawAmount(a.token, amounts.maxAmountsIn[i]),
             ),
         };
-    }
-
-    private getAmountsQuery(
-        poolTokens: Token[],
-        input: AddLiquidityInput,
-    ): AddLiquidityAmounts {
-        switch (input.kind) {
-            case AddLiquidityKind.Unbalanced: {
-                return {
-                    minimumBpt: 0n,
-                    maxAmountsIn: getAmounts(poolTokens, input.amountsIn),
-                    tokenInIndex: undefined,
-                };
-            }
-            case AddLiquidityKind.SingleToken: {
-                const tokenInIndex = poolTokens.findIndex((t) =>
-                    t.isSameAddress(input.tokenIn),
-                );
-                if (tokenInIndex === -1)
-                    throw Error("Can't find index of SingleToken");
-                const maxAmountsIn = Array(poolTokens.length).fill(0n);
-                maxAmountsIn[tokenInIndex] = MAX_UINT256;
-                return {
-                    minimumBpt: input.bptOut.rawAmount,
-                    maxAmountsIn,
-                    tokenInIndex,
-                };
-            }
-            case AddLiquidityKind.Proportional: {
-                return {
-                    minimumBpt: input.bptOut.rawAmount,
-                    maxAmountsIn: Array(poolTokens.length).fill(MAX_UINT256),
-                    tokenInIndex: undefined,
-                };
-            }
-        }
     }
 }
