+++ src/entities/removeLiquidity/removeLiquidityV2/composableStable/removeLiquidityComposableStable.ts
@@ -1,7 +1,7 @@
 import { encodeFunctionData } from 'viem';
 import { Token } from '../../../token';
 import { TokenAmount } from '../../../tokenAmount';
-import { VAULT, MAX_UINT256, ZERO_ADDRESS } from '../../../../utils/constants';
+import { VAULT, ZERO_ADDRESS } from '../../../../utils/constants';
 import { vaultV2Abi } from '../../../../abi';
 import { parseRemoveLiquidityArgs } from '../../../utils/parseRemoveLiquidityArgs';
 import {
@@ -12,11 +12,11 @@ import {
     RemoveLiquidityQueryOutput,
 } from '../../types';
 import { RemoveLiquidityV2ComposableStableBuildCallInput } from './types';
-import { RemoveLiquidityAmounts, PoolState } from '../../../types';
+import { PoolState } from '../../../types';
 import { doRemoveLiquidityQuery } from '../../../utils/doRemoveLiquidityQuery';
 import { ComposableStableEncoder } from '../../../encoders/composableStable';
-import { getAmounts, getSortedTokens } from '../../../utils';
-import { removeLiquiditySingleTokenExactInShouldHaveTokenOutIndexError } from '@/utils';
+import { getSortedTokens } from '../../../utils';
+import { getAmountsCall, getAmountsQuery } from '../../helper';
 
 export class RemoveLiquidityComposableStable implements RemoveLiquidityBase {
     public async query(
@@ -27,7 +27,7 @@ export class RemoveLiquidityComposableStable implements RemoveLiquidityBase {
         const bptIndex = poolState.tokens.findIndex(
             (t) => t.address === poolState.address,
         );
-        const amounts = this.getAmountsQuery(sortedTokens, input, bptIndex);
+        const amounts = getAmountsQuery(sortedTokens, input, bptIndex);
         const amountsWithoutBpt = {
             ...amounts,
             minAmountsOut: [
@@ -77,50 +77,10 @@ export class RemoveLiquidityComposableStable implements RemoveLiquidityBase {
         };
     }
 
-    private getAmountsQuery(
-        tokens: Token[],
-        input: RemoveLiquidityInput,
-        bptIndex: number,
-    ): RemoveLiquidityAmounts {
-        switch (input.kind) {
-            case RemoveLiquidityKind.Unbalanced:
-                return {
-                    minAmountsOut: getAmounts(tokens, input.amountsOut),
-                    tokenOutIndex: undefined,
-                    maxBptAmountIn: MAX_UINT256,
-                };
-            case RemoveLiquidityKind.SingleTokenExactOut:
-                return {
-                    minAmountsOut: getAmounts(tokens, [input.amountOut]),
-                    tokenOutIndex: tokens
-                        .filter((_, index) => index !== bptIndex)
-                        .findIndex((t) =>
-                            t.isSameAddress(input.amountOut.address),
-                        ),
-                    maxBptAmountIn: MAX_UINT256,
-                };
-            case RemoveLiquidityKind.SingleTokenExactIn:
-                return {
-                    minAmountsOut: Array(tokens.length).fill(0n),
-                    tokenOutIndex: tokens
-                        .filter((_, index) => index !== bptIndex)
-                        .findIndex((t) => t.isSameAddress(input.tokenOut)),
-                    maxBptAmountIn: input.bptIn.rawAmount,
-                };
-            case RemoveLiquidityKind.Recovery:
-            case RemoveLiquidityKind.Proportional:
-                return {
-                    minAmountsOut: Array(tokens.length).fill(0n),
-                    tokenOutIndex: undefined,
-                    maxBptAmountIn: input.bptIn.rawAmount,
-                };
-        }
-    }
-
     public buildCall(
         input: RemoveLiquidityV2ComposableStableBuildCallInput,
     ): RemoveLiquidityBuildCallOutput {
-        const amounts = this.getAmountsCall(input);
+        const amounts = getAmountsCall(input);
         const amountsWithoutBpt = {
             ...amounts,
             minAmountsOut: [
@@ -162,39 +122,5 @@ export class RemoveLiquidityComposableStable implements RemoveLiquidityBase {
                 TokenAmount.fromRawAmount(a.token, amounts.minAmountsOut[i]),
             ),
         };
-    }
-
-    private getAmountsCall(
-        input: RemoveLiquidityV2ComposableStableBuildCallInput,
-    ): RemoveLiquidityAmounts {
-        switch (input.removeLiquidityKind) {
-            case RemoveLiquidityKind.Unbalanced:
-            case RemoveLiquidityKind.SingleTokenExactOut:
-                return {
-                    minAmountsOut: input.amountsOut.map((a) => a.amount),
-                    tokenOutIndex: input.tokenOutIndex,
-                    maxBptAmountIn: input.slippage.applyTo(input.bptIn.amount),
-                };
-            case RemoveLiquidityKind.SingleTokenExactIn:
-                if (input.tokenOutIndex === undefined) {
-                    throw removeLiquiditySingleTokenExactInShouldHaveTokenOutIndexError;
-                }
-                return {
-                    minAmountsOut: input.amountsOut.map((a) =>
-                        input.slippage.applyTo(a.amount, -1),
-                    ),
-                    tokenOutIndex: input.tokenOutIndex,
-                    maxBptAmountIn: input.bptIn.amount,
-                };
-            case RemoveLiquidityKind.Recovery:
-            case RemoveLiquidityKind.Proportional:
-                return {
-                    minAmountsOut: input.amountsOut.map((a) =>
-                        input.slippage.applyTo(a.amount, -1),
-                    ),
-                    tokenOutIndex: input.tokenOutIndex,
-                    maxBptAmountIn: input.bptIn.amount,
-                };
-        }
     }
 }
